fix(screamingface): unify Engine origin presentation
Refs: OME-805
