+++ packages/screamingface/src/screamingface/_ui/connection_view.py
@@ -5,10 +5,8 @@ from __future__ import annotations
 from html import escape
 from typing import TYPE_CHECKING, Any, Protocol
 
-from screamingface._ui.connection_state import (
-    _ConnectionPanelState,
-    _is_screamingface_engine,
-)
+from screamingface._ui.connection_state import _ConnectionPanelState
+from screamingface._ui.engine_origin import _is_screamingface_engine
 from screamingface._ui.provider_icons import provider_icon_html
 from screamingface._ui.style import STYLE
